[Silver V] Title: 공포의 면담실, Time: 340 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/30088. 공포의 면담실/공포의 면담실.cc
@@ -3,7 +3,7 @@
 #include <algorithm>
 using namespace std;
 int main(){
-    int t,s=0;
+    long long t,s=0;
     cin >> t;
     vector<int> A(t);
     while(t--){
